split telmon into subpackage and entrypoint
